+++ hyp3_sdk/exceptions.py
@@ -4,15 +4,23 @@ from requests import Response
 from requests.exceptions import HTTPError
 
 
-class HyP3Error(Exception):
+class HyP3SDKError(Exception):
     """Base Exception for the HyP3 SDK"""
 
 
-class ASFSearchError(HyP3Error):
+class HyP3Error(HyP3SDKError):
+    """Raise for errors when using the HyP3 module"""
+
+
+class ASFSearchError(HyP3SDKError):
     """Raise for errors when using the ASF Search module"""
 
 
-class AuthenticationError(HyP3Error):
+class ServerError(HyP3SDKError):
+    """Raise when the HyP3 SDK encounters a server error"""
+
+
+class AuthenticationError(HyP3SDKError):
     """Raise when authentication does not succeed"""
 
 
@@ -22,7 +30,7 @@ def _raise_for_hyp3_status(response: Response):
     except HTTPError:
         if 400 <= response.status_code < 500:
             raise HyP3Error(f'{response} {response.json()["detail"]}')
-        raise
+        raise ServerError
 
 
 def _raise_for_search_status(response: Response):
@@ -31,4 +39,4 @@ def _raise_for_search_status(response: Response):
     except HTTPError:
         if 400 <= response.status_code < 500:
             raise ASFSearchError(f'{response} {response.json()["error"]["report"]}')
-        raise
+        raise ServerError